kernel32: "Fixed" enumerating 64-bit process modules
For whatever reason, we get errno 299 (ERROR_PARTIAL_COPY) when
calling EnumProcessModulesEx on a 64-bit process. I cannot find
any documentation or explanation as to why this happens other
than this blog post that discusses a 32-bit application doing
the same thing to a 64-bit application:

https://winprogger.com/getmodulefilenameex-enumprocessmodulesex-failures-in-wow64/

It looks like we still get the full list of modules. I compared
the number of modules returned using the different dwFilterFlag
values, and they were all the same.

This commit also implements requesting the number of modules
loaded in a process, allowing us to remove the hard-coded
limit we had in place. Ironically, that limit comes from
a Windows documentation example. Make of that what you will.