plugins: Share memory using plugin's allocator
This commit almost completely reworks how memshonk and plugins
share memory. Now, memshonk will almost always pass memory by
allocating a shared buffer object. This object contains the size
of the buffer, excluding the bits that contain the size. I.e.,
a buffer of 32 bytes is actually 36 bytes in size with the first
4 bytes being used to store its usable size as an unsigned 32-bit
integer.

In addition, I restructured how the memshonk callbacks are managed.
The new strategy allows the callback functions to "know" which plugin
executed them. As a result, the memshonk callback code can create
shared buffers using the plugin's memory allocator and let the plugin
take ownership of that memory. This change was necessary to make the
new shared memory strategy feasible.