plugins: Separate Ctx into Ctx and CtxCloser
This is a safer implementation that also fixes a Rust panic on Windows
in the mpsc library. I am not sure why the panic was only occurring on
Windows. My theory is we were deallocating the Ctx before Rust was
finished with it.

The new design avoids such a scenario by separating the Ctx object
into two parts: one owned by memshonk and one owned by the plugin.
Deallocating one does not affect the other.