progctl: Renamed OS-specific process types
The current naming scheme has been driving us crazy ("oh, that is
in unixProcess in process_unix"). This commit makes the type names
consistent with the source file names. I think we would prefer
"osNameProcess" ordinarily, but the file naming is less flexible
in this context... so that is what we will follow.